Solved the MA and EMA
I finally am able to solve for the MA and also get the EMA. I am using a formula and just getting the prices by myself through the API and calculating the technical indicators. I will want to see whether the numbers are correct and if I would be making a profit according to my program. Next step would be to gather data from different days and compare how my algorithm would work if I change certain parameters and which would give a more consistent and higher return